pyproject.toml
[tool.poetry]
name = "rc-sdk"
version = "0.1.0"
description = "RingCentral Call Center SDK"
authors = ["Cloud Warriors"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
pydantic = "^2.7.1"
redis = "^5.0.4"
prometheus-client = "^0.20.0"
python-json-logger = "^2.0.7"
tenacity = "^8.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
